281: create new /control file structure
